+++ src/renderer/components/atlas/Game.tsx
@@ -24,6 +24,7 @@ import { mapMiniPic } from '../../lib/Resources';
 import Player from './Player';
 import { CharacterIcon } from './CharacterIcon';
 import { useTranslation } from 'react-i18next';
+import { trackEvent } from '@aptabase/electron/renderer';
 
 export const TeamFlexBox = styled(FlexBox)(() => ({
   paddingLeft: 20,
@@ -130,7 +131,10 @@ export default function Game({ info, playerData, expanded }: Props) {
         <Tooltip title={`${t(`maps.${info.map}`)} ${info.ts}`} arrow>
           <Box
             flexBasis={120}
-            onClick={() => setCollapsed((x) => !x)}
+            onClick={() => {
+              trackEvent('Game Details');
+              setCollapsed((x) => !x);
+            }}
             style={{
               backgroundImage: `url(${mapMiniPic(info.map)})`,
               backgroundSize: 'cover',