+++ src/renderer/components/pages/SettingsPage.tsx
@@ -377,15 +377,9 @@ export default function SettingsPage() {
             <FormControl fullWidth>
               <Select value={ip} onChange={handleChange}>
                 <MenuItem value="evos-emu.com">{t('ips.noProxy')}</MenuItem>
-                <MenuItem value="arproxy.addalyn.baby">
-                  {t('ips.proxy1')}
-                </MenuItem>
-                <MenuItem value="arproxy2.addalyn.baby">
-                  {t('ips.proxy2')}
-                </MenuItem>
-                <MenuItem value="arproxy3.addalyn.baby">
-                  {t('ips.proxy3')}
-                </MenuItem>
+                <MenuItem value="de.evos.live">{t('ips.proxy1')}</MenuItem>
+                <MenuItem value="fr.evos.live">{t('ips.proxy2')}</MenuItem>
+                <MenuItem value="fi.evos.live">{t('ips.proxy3')}</MenuItem>
               </Select>
             </FormControl>
           </Grid>
@@ -440,7 +434,8 @@ export default function SettingsPage() {
               <FormControlLabel
                 control={<Switch />}
                 label={t('settings.autoPatchingLabel')}
-                checked={enablePatching === 'true'}
+                checked={false} // {enablePatching === 'true'}
+                disabled={true} // Only for christmas we need it next year!
                 onChange={() => {
                   setEnablePatching(
                     enablePatching === 'true' ? 'false' : 'true',